Cleaned up spike generation
There was a problem in initialising spike's activiity as false and then removing them immediately if they are. That was the root of the problem otherwise its fine. This should be the final version of the code that Im going to submit unless i have time to implement further modifications like the death animation